refactor: MediaModal to accept search term as a prop
This commit refactors the MediaModal so that the image/file search
term is a prop instead of a state. The reason this needs to be done
is because the MediaSettingsModal is on the same hierarchy level as
the MediaModal - when we upload a new file/image, we do it through
the MediaSettingsModal, whose parent component is the EditPage or
other layout page, not the MediaModal.

This means we need to pass the state to the MediaModal as a prop from
the shared parent component that MediaModal has with MediaSettingsModal.
(If MediaSettingsModal was a child component of MediaModal, we would be
able to define a function which alters the state of the search term inside
MediaModal from MediaSettingsModal and pass it as a prop from MediaModal
to MediaSettingsModal).