feat: filter for newly uploaded image
This commit updates the EditPage component so that when we upload a
new image using the MediaSettingsModal, when the MediaModal is
rendered again, we pre-fill the search bar with the file name of the
recently uploaded file so that users do not have to scroll through all
their images to find it. This is because we assume that users are only
uploading an image in the EditPage component because they want to use it
immediately, which is why we should make it readily available for
selection in the MediaModal.