fix: disable search bar pre-fill feature
This commit disables the feature where the search bar is prefilled upon
uploading an image. We also prevent the image modal from reopening,
as we work towards a state where upon upload, the image is automatically
selected/inserted instead of having to select it again after upload.

The actual feature where the image is automatically selected/inserted
will be implemented in a subsequent commit.