feat: automatically select/insert image on upload
Previously, the media settings modal would pre-fill the search bar
after an image is uploaded to allow users to select the newly uploaded
images. However, this was decided to be clunky - it would be preferable
for us to automatically select the image / insert the image into the
editor upon upload. This commit updates the upload functionality so
that the image is automatically selected/inserted upon upload.