fix: restrict access to login/logout state setting
Currently, we allow the login state to be set to logged out repeatedly,
even if the logout state is already logged out. This affects one of the
useEffect blocks which depends on `isLoggedIn` in its dependency array.

This commit makes the fix to only allow the `isLoggedIn` hook to be
set to false when `isLoggedIn` is true.