Feat/media modal search bar (#313)
* fix: file loader bug

Steps to recreate bug:
1. Go to the Files or Images layout
2. Click on upload file
3. Select a file. Now the file naming modal will pop up
4. Exit from the modal
5. Click on "Upload File" again
6. This time, instead of selecting a file, click cancel

This time, an error will pop up as event.target.files[0] is undefined.
This is because the browser file list is now empty.

This fix checks that there exists a file in the fileList before
attempting to read the file for upload.

* fix: unable to select same file twice for upload

Steps to reproduce bug:
1. Go to the Files layout
2. Click on upload file
3. Select a file. At this point, the file renaming modal should pop up.
4. Exit the file renaming modal
5. Click on upload file again
6. Select the same file
7. Nothing happens - the file renaming modal does not show up

The reason this bug occurs is that the onload event that our filereader
detects for changes to the file input value. When there is no change
to the value, the `fileReader.onload` event doesn't get triggered.

The fix is to reset the file input value after every use.

For some reason, this bug does not affect image uploads.

* fix: MediaSettingsModal expects boolean prop instead of string

This commit fixes a problem where we set a string value for
the isPendingImageUpload prop for the MediaSettingsModal. The component
expects a boolean prop instead.

* feat: add MediaSearchBar component and styles

This commit adds a search bar component for the MediaModal component
as well as the corresponding styles for it. The search bar filters images
in the media modal by file name for the images/files which contain
the search term as a substring.

* feat: implement media modal search

This commit implements the search bar for the MediaModal component
by filtering out existing media files by file name given a search term

* fix: remove height specification for content cards

This commit removes the height specification for content cards, which
was artificially suppressing the height of the "Upload Image"/"Upload File"
card in the MediaModal.

* refactor: MediaModal to accept search term as a prop

This commit refactors the MediaModal so that the image/file search
term is a prop instead of a state. The reason this needs to be done
is because the MediaSettingsModal is on the same hierarchy level as
the MediaModal - when we upload a new file/image, we do it through
the MediaSettingsModal, whose parent component is the EditPage or
other layout page, not the MediaModal.

This means we need to pass the state to the MediaModal as a prop from
the shared parent component that MediaModal has with MediaSettingsModal.
(If MediaSettingsModal was a child component of MediaModal, we would be
able to define a function which alters the state of the search term inside
MediaModal from MediaSettingsModal and pass it as a prop from MediaModal
to MediaSettingsModal).

* feat: filter for newly uploaded image

This commit updates the EditPage component so that when we upload a
new image using the MediaSettingsModal, when the MediaModal is
rendered again, we pre-fill the search bar with the file name of the
recently uploaded file so that users do not have to scroll through all
their images to find it. This is because we assume that users are only
uploading an image in the EditPage component because they want to use it
immediately, which is why we should make it readily available for
selection in the MediaModal.

* feat: filter for newly uploaded image with FormFieldMedia

This commit makes the same update from the previous commit (that when
we upload a new image using the MediaSettingsModal, we prefill the search
bar in the MediaModal with the file name of the recently uploaded file) but
on the FormFieldMedia component, which is the only other place where the
MediaModal is used.

* fix: set width of media cards container in MediaModal to be 100%

* fix: disable search bar pre-fill feature

This commit disables the feature where the search bar is prefilled upon
uploading an image. We also prevent the image modal from reopening,
as we work towards a state where upon upload, the image is automatically
selected/inserted instead of having to select it again after upload.

The actual feature where the image is automatically selected/inserted
will be implemented in a subsequent commit.

* feat: update media modal upload button

This commit updates the `SaveDeleteButtons` component so that it
accepts `saveLabel` and `deleteLabel` props to override the default
button labels.

It also updates the media settings modal "Save" button to read "Upload"
for clarity

* chore: correct typo

* feat: automatically select/insert image on upload

Previously, the media settings modal would pre-fill the search bar
after an image is uploaded to allow users to select the newly uploaded
images. However, this was decided to be clunky - it would be preferable
for us to automatically select the image / insert the image into the
editor upon upload. This commit updates the upload functionality so
that the image is automatically selected/inserted upon upload.

* fix: replace imageSearchTerm with mediaSearchTerm

Co-authored-by: Jie Hao Kwa <jiehaokwa@142.121.96.27.static.m1net.com.sg>
Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>
Co-authored-by: jiehao <jiehao@open.gov.sg>