fix: track axios errors synchronously
The main problem we're trying to solve is that the useEffect containing
the logout warning alert is triggered multiple times because the axios
interceptor function, which triggers the original useEffect, was being
called multiple times.

This is related to the use of state hooks in the interceptor function
- React is unable to process the state updates quickly enough as the
interceptor function is called multiple times synchronously (although
I'm not clear on the exact reasoning behind this yet).

This can be solved by using a synchronous variable instead of the React
state to keep track of the axios errors, to prevent multiple triggers of
the axios error interceptor function in the event of a 401 error. This
in turn, prevents the multiple triggers of the original useEffect which
triggers the logout warning alert.