fix: cancel axios requests if page is dismounted
The previous attempt to solve the auto-logout problem tried to keep
track of the number of `401` errors our app had received, and to only
trigger the autologout flow for the first `401` error. This turned out
to be problematic because there is something wrong in our component
definition which causes a re-render of the root `App` component. This
causes the axios error count to restart from 0, which means that we still
get multiple triggerings of the auto-logout flow.

There are a couple of ways to tackle this problem:
1. figure out the logical problem which causes the unnecessary re-renders
2. prevent the auto-logout flow from being triggered by cancelling axios
requests upon page dismount

Despite several attempts at 1, I was unable to uncover the logical
inconsistency behind our re-renders, so I decided to try cancelling the
axios requests instead. Cancelling the requests will prevent the error
interceptor function from being called, and thus will prevent our autologout
flow from being triggered.

This commit starts by using the `Workspace` component as an experiment.
Subsequent commits will extend the use of the cancel token to other
components.