fix: remove axios cancel requests
In the previous commits, I talked about how cancelling axios requests
would help us to prevent triggering the auto-logout flow. However, there
were some flaws in the reasoning for that approach.

There are a couple of ways that approach would fail.
1. If the timing of the network calls were such that all the requests were
cancelled and none of the requests returned a 401, the autologout flow would
not be triggered at all.
2. It is still possible for multiple 401s to be returned

Subsequently, I will focus on preventing re-renders so the autologout flow
does not get triggered multiple times