fix: coordinate state updates using boolean checks
The root `App` component was being re-rendered due to state updates
on its state hooks. This was because we did not properly check for
state hook values before continuing to update hook values. Since
react state updates asynchronously, the hooks were being updated
in a manner that was unpredictable before.

This commit adds the relevant boolean checks to prevent the
useEffect blocks from being run unnecessarily.