Fix/double auto logout (#336)
* fix: restrict access to login/logout state setting

Currently, we allow the login state to be set to logged out repeatedly,
even if the logout state is already logged out. This affects one of the
useEffect blocks which depends on `isLoggedIn` in its dependency array.

This commit makes the fix to only allow the `isLoggedIn` hook to be
set to false when `isLoggedIn` is true.

* fix: prevent state update on unmount

This commit updates the ProtectedRoute component to prevent state updates
on unmount

* fix: cleanup protected route logic

* fix: track axios errors synchronously

The main problem we're trying to solve is that the useEffect containing
the logout warning alert is triggered multiple times because the axios
interceptor function, which triggers the original useEffect, was being
called multiple times.

This is related to the use of state hooks in the interceptor function
- React is unable to process the state updates quickly enough as the
interceptor function is called multiple times synchronously (although
I'm not clear on the exact reasoning behind this yet).

This can be solved by using a synchronous variable instead of the React
state to keep track of the axios errors, to prevent multiple triggers of
the axios error interceptor function in the event of a 401 error. This
in turn, prevents the multiple triggers of the original useEffect which
triggers the logout warning alert.

* fix: prevent state update while unmounting

* chore: clean up unused code

* fix: cancel axios requests if page is dismounted

The previous attempt to solve the auto-logout problem tried to keep
track of the number of `401` errors our app had received, and to only
trigger the autologout flow for the first `401` error. This turned out
to be problematic because there is something wrong in our component
definition which causes a re-render of the root `App` component. This
causes the axios error count to restart from 0, which means that we still
get multiple triggerings of the auto-logout flow.

There are a couple of ways to tackle this problem:
1. figure out the logical problem which causes the unnecessary re-renders
2. prevent the auto-logout flow from being triggered by cancelling axios
requests upon page dismount

Despite several attempts at 1, I was unable to uncover the logical
inconsistency behind our re-renders, so I decided to try cancelling the
axios requests instead. Cancelling the requests will prevent the error
interceptor function from being called, and thus will prevent our autologout
flow from being triggered.

This commit starts by using the `Workspace` component as an experiment.
Subsequent commits will extend the use of the cancel token to other
components.

* fix: remove axios cancel requests

In the previous commits, I talked about how cancelling axios requests
would help us to prevent triggering the auto-logout flow. However, there
were some flaws in the reasoning for that approach.

There are a couple of ways that approach would fail.
1. If the timing of the network calls were such that all the requests were
cancelled and none of the requests returned a 401, the autologout flow would
not be triggered at all.
2. It is still possible for multiple 401s to be returned

Subsequently, I will focus on preventing re-renders so the autologout flow
does not get triggered multiple times

* refactor: access login state from context instead of props

* fix: coordinate state updates using boolean checks

The root `App` component was being re-rendered due to state updates
on its state hooks. This was because we did not properly check for
state hook values before continuing to update hook values. Since
react state updates asynchronously, the hooks were being updated
in a manner that was unpredictable before.

This commit adds the relevant boolean checks to prevent the
useEffect blocks from being run unnecessarily.

Co-authored-by: jiehao <jiehao@open.gov.sg>