Fix: re-enabling save button after fixing error
Previously, we encountered an issue with the save button not returning after an error had been corrected. This was due to validateSections modifying the error itself, while our useEffect checked if the errors object had been modified - as a result, this useEffect never fired.

This commit resolves this by cloning the error section before before validate it.