fix: eliminate additional re-render of component when retrieving colors
This commit eliminates the additional re-render that occurs when
one of our layout components attempts to retrieve a repo's site colors
by eliminating the use of the React `Context` and `useState` API.

Currently, we store site color as a state variable at the root app
level. We then pass it into a context for it to be consumed by
all of the root app's children. When we enter a new site's workspace,
we retrieve the site's site colors, which updates state at the root app
level. This triggers a re-render of the root app, which subsequently
triggers a re-render of its children. Even though we don't pass the
color variables as props, the components are rerendered because we
use the `render` prop in our `ProtectedRoute` HOC.

To prevent this sort of re-rendering, we store and retrieve the site colors
in local storage. This leads to a much smoother user experience.