feat: implement Folders layout for subfolder paths
This commit implements the Folders layout for subfolder paths by
checking whether the path references a subfolder. If it does, we
parse the directory file using a different util function which
filters out all files that do not belong to the subfolder, and
render only those files within the subfolder.

However, there remains a bug which needs to be fixed - when defining
the behavior for the back button in the Header, we want to go back
to the workspace if we are in a folder, and back to the parent folder
if we are in a subfolder. However, currently when we are in a subfolder,
the back button transports the user all the way back to the Workspace.