Feat/update site colors without re-render (#340)
* feat: create context for storing site colors

This commit creates a React context for storing, accessing, and
setting site colors so that we don't have to rely on prop-drilling.

* feat: create custom hook for site color retrieval logic

This commit creates a custom hook which encapsulates the state and
the logic for site color retrieval and CSS stylesheet generation so
that we don't need to repeat ourselves in each component we use it in.

* refactor: update layouts to use SiteColorsContext

This commit refactors the following layouts which currently use
site colors:
- EditHomepage
- EditContactUs
- EditPage

Post-refactor, they now use the methods passed by the SiteColorsContext
to retrieve site colors and to generate custom style sheets for each
repo

* fix: eliminate additional re-render of component when retrieving colors

This commit eliminates the additional re-render that occurs when
one of our layout components attempts to retrieve a repo's site colors
by eliminating the use of the React `Context` and `useState` API.

Currently, we store site color as a state variable at the root app
level. We then pass it into a context for it to be consumed by
all of the root app's children. When we enter a new site's workspace,
we retrieve the site's site colors, which updates state at the root app
level. This triggers a re-render of the root app, which subsequently
triggers a re-render of its children. Even though we don't pass the
color variables as props, the components are rerendered because we
use the `render` prop in our `ProtectedRoute` HOC.

To prevent this sort of re-rendering, we store and retrieve the site colors
in local storage. This leads to a much smoother user experience.

* feat: cleanup site colors on app start

This commit resets the LOCAL_STORAGE_SITE_COLORS variable on app
start so that updated site colors are retrieved each time.

Co-authored-by: jiehao <jiehao@open.gov.sg>