fix: reset `Header` state after redirection
Previously, we encountered a bug where in a subfolder, when we
attempted to click the back button in the Header component to go
back to the parent folder, we ended up redirecting all the way back
to the Workspace layout. The reason for this is because we are using
the same Folder layout to render both folders and subfolders, hence
the Header component does not unmount. This means that the
`shouldPerformBackNav` state remains true and automatically performs
the back nav redirection to the Workspace upon loading the parent folder.

The solution to this is to reset our `Header` state. Since the
`shouldRedirect` and `shouldPerformBackNav` states should only be
`true` ephemerally, we can set up a useEffect which sets them back
to false once they are set to true. This would have the effect of
resetting the `Header` component state when the user is redirected
back to the parent folder.