feat: disable creation of new sub-folders within sub-folders
This commit adds the necessary logic and styles to disable the
"Create new sub-folders" button in the Folders layout when the
Folders layout is rendering a sub-folder already. As discussed,
Isomer does not accommodate nesting in collections/folders beyond
a single level, so we should not allow creation of sub-folders
within sub-folders.