Feat/new folder layout (#350)
* feat: add skeleton for folders layout

This commit adds the layout file and the necessary styles needed for
the folders layout and the three option buttons stipulated in the
design.

* feat: add folder content container and items and associated styles

This commit adds the FolderContent component, which is a container
which stores all the folder's items (files or directories), each of
which is a FolderContentItem. It also introduces the necessary styles
for these components.

This commit also makes minor edits to the styles introduced in the
previous commit.

* feat: add hover state for folder option buttons

* feat: add frontend routes for Folders layout

This commit adds the /sites/:siteName/folder/:folderName and
/sites/:siteName/folder/:folderName/:subfolderName routes to our App,
and both routes will hit our Folders layout.

This commit also adds minor functionality - in particular, it will
toggle the "Rearrange items" button when that button is clicked

* feat: install react-query

This commit installs react-query into our app, as well as react-query
devtools for local development

* fix: place react-query devtools only in Folders layout

This commit places the react-query devtools component only in the
Folders layout because that is the only component that uses react-query
for now

* feat: create api file for storing data fetching and mutation functions

* feat: helper function to parse collection.yml directory file

This helper function parses the collection.yml directory file and
outputs an array which specifies the order in which the pages are
sorted in the collection.

A collection page will have the following attributes:
- `type` as `file`
- `path` as the file path
- `title` as the deslugified file path with the file extension stripped

A subfolder entry will have the following attributes:
- `type` as `dir`
- `title` as the subfolder title (for example for the entry `prize-jury/test.md`
the title will be `prize-jury`)
- `children` as an array of collection page entries, which have the
attributes explained above

* feat: first cut of Folder layout

This commit adds the necessary changes to the `FolderContent` component
so that each item in the folder (file or subfolder) is represented
with a long horizontal bar.

Note that there is still no functionality (the folder options buttons
still do not work, the 3 dots on the individual item bars will also
not do anything at this point in time)

* feat: implement correct header back button behavior for Folder layout

If the Folder layout is for a subfolder, the back button should be
to the parent folder.

If the Folder layout is for a folder, the back
button should be to the Workspace.

* fix: modify folder content items to navigate to EditPage

This commit updates the FolderContentItem component so that if it's
a directory, it links to the subfolder and if it's a file, it links
to the EditPage layout

* feat: implement Folders layout for subfolder paths

This commit implements the Folders layout for subfolder paths by
checking whether the path references a subfolder. If it does, we
parse the directory file using a different util function which
filters out all files that do not belong to the subfolder, and
render only those files within the subfolder.

However, there remains a bug which needs to be fixed - when defining
the behavior for the back button in the Header, we want to go back
to the workspace if we are in a folder, and back to the parent folder
if we are in a subfolder. However, currently when we are in a subfolder,
the back button transports the user all the way back to the Workspace.

* fix: reset `Header` state after redirection

Previously, we encountered a bug where in a subfolder, when we
attempted to click the back button in the Header component to go
back to the parent folder, we ended up redirecting all the way back
to the Workspace layout. The reason for this is because we are using
the same Folder layout to render both folders and subfolders, hence
the Header component does not unmount. This means that the
`shouldPerformBackNav` state remains true and automatically performs
the back nav redirection to the Workspace upon loading the parent folder.

The solution to this is to reset our `Header` state. Since the
`shouldRedirect` and `shouldPerformBackNav` states should only be
`true` ephemerally, we can set up a useEffect which sets them back
to false once they are set to true. This would have the effect of
resetting the `Header` component state when the user is redirected
back to the parent folder.

* feat: disable creation of new sub-folders within sub-folders

This commit adds the necessary logic and styles to disable the
"Create new sub-folders" button in the Folders layout when the
Folders layout is rendering a sub-folder already. As discussed,
Isomer does not accommodate nesting in collections/folders beyond
a single level, so we should not allow creation of sub-folders
within sub-folders.

* feat: error-handling for folders layout

This commit ensures consistency in behavior between the new Folders
layout and the old CollectionPagesSection component. If a user enters
a URL which leads to an undefined folder/folder does not exist, then
we redirect them to the Workspace.

* fix: bug in directory file parsing logic

* feat: make Folder breadcrumb navigable

This commit makes a change to the breadcrumb in the Folders layout
so that if a user is in a subfolder, the breadcrumb header representing
the parent folder can be clicked on, and links to the Folder layout
which renders the parent folder.

* fix: re-enable window refocusing fetching for react-query

* feat: error-handling for invalid subfolder names

If an invalid subfolder name is passed to the URL as a path param,
we redirect the user to the workspace.

Co-authored-by: jiehao <jiehao@open.gov.sg>