feat: create LoadingButton component compatible with react-query
By using react-query, we are also provided with the loading state
of our api call - this means we don't need to handle it ourselves
within our LoadingButton component. This commit introduces a simplified
loading button which simply receives `isLoading` as a prop and renders
a loading spinner and loading styles when `isLoading` is true.