feat: update EditNavBar GET request to retrieve collection/folder data
Previously, we used the GET /v1/sites/:siteName/collections endpoint
to retrieve data on collections. However, this endpoint is outdated -
it retrieves data on collections based on the _config.yml file. This
won't work with our Jekyll 4.2 repos going forward, since information
on collection data will be stored in each individual collection directory.

This commit updates the GET request to retrieve from the GET
v1/sites/:siteName/folders/all endpoint instead. This endpoint returns
not just information on what collections exist in the repo, but also
returns information on the order of pages within each collection, which
we need for rendering the collection page dropdown in the NavBar template.