refactor: use react-query for EditNavBar (#358)
* chore: import react-query and ReactQueryDevTools

* chore: cleanup code

* fix: define frontend route for EditNavBar

* refactor: EditNavBar get data to use react-query

This commit refactors the `EditNavBar` component to use react-query
to retrieve nav bar data from the backend. This includes moving the
API call into a callback function defined in the api.js file.

* feat: create LoadingButton component compatible with react-query

By using react-query, we are also provided with the loading state
of our api call - this means we don't need to handle it ourselves
within our LoadingButton component. This commit introduces a simplified
loading button which simply receives `isLoading` as a prop and renders
a loading spinner and loading styles when `isLoading` is true.

* refactor: use react-query to update navigation data

* chore: cleanup code

Co-authored-by: jiehao <jiehao@open.gov.sg>