feat: add success toast after page-reordering
This commit adds a green success toast after a folder page order has
been successfully updated.

This commit also refactors the GET and mutation queries so that the
useQuery callback functions are responsible for error-handling as well.
This makes the callback functions more self-contained, so that we don't
need to manage query outcome and accompanying behavior in two separate
places (Folders layout and api file) - now we handle it all in the
api file.