Feat/display collection in nav bar (#362)
* chore: import react-query and ReactQueryDevTools

* refactor: EditNavBar get data to use react-query

This commit refactors the `EditNavBar` component to use react-query
to retrieve nav bar data from the backend. This includes moving the
API call into a callback function defined in the api.js file.

* chore: correct react keys for NavBar template

* feat: update EditNavBar GET request to retrieve collection/folder data

Previously, we used the GET /v1/sites/:siteName/collections endpoint
to retrieve data on collections. However, this endpoint is outdated -
it retrieves data on collections based on the _config.yml file. This
won't work with our Jekyll 4.2 repos going forward, since information
on collection data will be stored in each individual collection directory.

This commit updates the GET request to retrieve from the GET
v1/sites/:siteName/folders/all endpoint instead. This endpoint returns
not just information on what collections exist in the repo, but also
returns information on the order of pages within each collection, which
we need for rendering the collection page dropdown in the NavBar template.

* chore: remove log statement

* feat: get nav collection dropdown content from folder order

This commit adds a util function which parses the folder order for
each folder/collection and returns them in an array of deslugified
elements in accordance with the nav bar dropdown requirements.

* feat: update EditNavBar to display collection/folder dropdown

This commit updates the EditNavBar layout with a new folderDropdowns
state variable, which is passed to the TemplateNavBar as the
collectionsInfo prop. This gives the TemplateNavBar component the
necessary information to render the dropdown items for collections/folders

* fix: remove unnecessary slice since folder name now starts WITHOUT _

* fix: folder comparison when constructing EditNavBar dropdown

Co-authored-by: jiehao <jiehao@open.gov.sg>