fix: Image routing for folders nested beyond 2nd level
The frontend paths that we use for Images are:
- /sites/:siteName/images
- /sites/:siteName/images/:customPath

The second path, /sites/:siteName/images/:customPath, is responsible
for handling all other nested levels. For example, if the repo path to
the nested image folder is /images/features/2013, then the CMS URL
would be /sites/:siteName/images/features/2013. The problem with this
is that this URL does not actually match the allocated URL, since
there are `/` characters in the route param. Therefore, we need to
modify the `customPath` variables to be encoded URI components so that
the accessed URLs match the frontend route.