fix: prevent multiple toasts from appearing on any given page
When our page as an error (for example, due to network error), we
trigger an error toast. When we move to a different tab/window and
back to the CMS tab, this error toast is triggered again since we
enabled window-focus refetching for react-query. When repeated
multiple times, we end up with a bunch of the same toasts on the same
page.

This PR adds a toast ID to each of the success and error toasts so that
we only have one toast of each type on screen at a time. This means we
prevent multiple error toasts, but still allow for an error toast, and then
a success toast to be triggered. See https://fkhadra.github.io/react-toastify/prevent-duplicate