Feat/create success and error toast wrapper functions (#364)
* feat: add success toast after page-reordering

This commit adds a green success toast after a folder page order has
been successfully updated.

This commit also refactors the GET and mutation queries so that the
useQuery callback functions are responsible for error-handling as well.
This makes the callback functions more self-contained, so that we don't
need to manage query outcome and accompanying behavior in two separate
places (Folders layout and api file) - now we handle it all in the
api file.

* fix: add React key for Folder draggables

* fix: update DEFAULT_ERROR_TOAST_MESSAGE to be more generic

* feat: create reusable toast wrapper functions

* fix: use RETRY_MSG to compose DEFAULT_ERROR_TOAST_MSG

This commit splits the DEFAULT_ERROR_TOAST_MSG constant in the utils
file to use a RETRY_MSG as well, since the default retry message is used
in many different places.

* chore: rename RETRY_MSG to DEFAULT_RETRY_MSG

* fix: update api file to use successToast and errorToast functions

* fix: update Folders layout to use onSuccess and onError config options

This commit updates the queries made in the Fodlers layout to use the
onSuccess and onError config options for react-query to handle calling
success and error toasts, simplifying the API functions.

* fix: update EditNavBar to use onSuccess and onError config options to call toast functions

* fix: update components to use new errorToast and successToast functions

This commit reduces code duplication by converting all existing
invocations of the toast functions to use the wrapped errorToast and
successToast functions we created a few commits ago.

* chore: add return statements to api functions

* fix: prevent multiple toasts from appearing on any given page

When our page as an error (for example, due to network error), we
trigger an error toast. When we move to a different tab/window and
back to the CMS tab, this error toast is triggered again since we
enabled window-focus refetching for react-query. When repeated
multiple times, we end up with a bunch of the same toasts on the same
page.

This PR adds a toast ID to each of the success and error toasts so that
we only have one toast of each type on screen at a time. This means we
prevent multiple error toasts, but still allow for an error toast, and then
a success toast to be triggered. See https://fkhadra.github.io/react-toastify/prevent-duplicate

* fix: use new useRedirectHook

Co-authored-by: jiehao <jiehao@open.gov.sg>