Feat/save new folder functionality (#369)
* Feat: add file saving functionality

* Refactor: use react-query to save

* Fix: duplicate check for folder names

* Fix: refetch subfolder data when url is changed

This commit enforces a refetch when the url changes. Previously, as Folder and Subfolder used the same component, attempting to redirect to subfolder from folder resulted in the folder contents not being refetched. This commit fixes this to enable fresh data when creating a new subfolder.

* Fix: use onError for mutation instead of separate useEffect

* Feat: add check for protected folder names

* Fix: change toast to use new errorToast util function

* Fix: redirect to workspace instead of not found

This commit makes the behaviour when the folder is not found the same as that when the subfolder is not found.

* Nit: remove unnecessary log statement

* Fix: remove original order option for sort

* Refactor: move duplicate and protected name check into validators

* Feat: add more details about protected folder names to error message

* Fix: use LoadingButton instead of LoadingButtonReactQuery

* Fix: button text color when selected

* Fix: remove unused isLoading

* Fix: update error message

* Fix: add return to await statement