fix: css-loader file resolution bug introduced by CRA v4
In recent commits, we upgraded our react-scripts version from 3.4.4
to 4.0.3. This is because CRA (create-react-app) v3 uses an outdated
version of eslint (facebook/create-react-app#8849). This introduced
a bug related to the css-loader library, which can no longer resolve
assets in the public folder:
- https://github.com/facebook/create-react-app/issues/9870#issuecomment-719823931
- https://github.com/webpack-contrib/css-loader/issues/1136#issuecomment-664984703

This commit fixes this bug by moving the referenced image to the
relevant sub-directory in the src directory.