refactor: place FolderModal component at the Folders layout for consistency
This commit moves the FolderModal component from the FolderContentItem
component to the Folders layout for consistency, since we also place the
DeleteWarningModal the the Folders layout and drill the props down to
FolderContentItem. Whether this is the desired distribution can be discussed
later - we can refactor this easily. The immediate concern is to preserve
consistency of behavior among the dropdown components.