refactor: simplify OverviewCard move file logic
Thanks to PR #124 on the backend https://github.com/isomerpages/isomercms-backend/pull/124,
we can simplify the file-moving logic in `OverviewCard` since `OverviewCard` is
now used only for unlinked pages in the workspace (it is also
used for Resource Category display, but these do not have any move file
functionality).

This allows us to remove all the complex logic associated with file-moving
and replace it with a single API call to the new file moving
endpoint. This also allows us to delete the now-unused util function
`saveFileAndRetrieveUrl`.