Feat/delete rename subfolders (#383)
* feat: add api functions for deleting and renaming subfolders

* refactor: FolderModal to use react-query for mutations

* feat: add subfolder renaming functionality to FolderModal

Previously, the FolderModal component catered to the Folder and
Resource category use cases. This commit expands the FolderModal component's
functionality so that it can be used to rename subfolders as well.
This was done by adding a new `subfolderName` prop, which is used to
populate the FormField if present.

* fix: correct variable name in api selection function

* fix: move FolderModal outside of Link component

With the FolderModal within the Link component, the Link component
is "clicked" whenever the FolderModal is invoked. Moving the FolderModal
component out restores normal behavior.

* refactor: place FolderModal component at the Folders layout for consistency

This commit moves the FolderModal component from the FolderContentItem
component to the Folders layout for consistency, since we also place the
DeleteWarningModal the the Folders layout and drill the props down to
FolderContentItem. Whether this is the desired distribution can be discussed
later - we can refactor this easily. The immediate concern is to preserve
consistency of behavior among the dropdown components.

* chore: rename api function

* feat: update Folders layout to allow for deletion of subfolders

* feat: automatically refetch folder contents after deletion of page/subfolder

* fix: slugify new folder, subfolder, category names

* fix: selected item type bug

* chore: remove unused test files

Co-authored-by: jiehao <jiehao@open.gov.sg>